Wire up UI to service for general navigation flow
This change wires up the existing Angular-based UI to the new service so
that the user can navigate from a noun through an item from that noun
and then on to the property page for a verb of that noun.  Verbs are not
executable at this moment, that will be coming next.