Introduce Electron as desktop shell for the web UI
This change adds the Electron desktop application shell so that the
Phosphor web UI can be hosted there by default.  This provides a more
desktop-like experience when launching Show-Module from the command
line.  An -OpenInBrowser parameter was also added to Show-Module to
allow opening the web UI in a browser in cases where that is desirable.